// sdks/node-ts/src/apache_beam/proto/beam_expansion_api.ts
export interface ExpansionRequest {
    /**
     * Set of components needed to interpret the transform, or which
     * may be useful for its expansion.  This includes the input
     * PCollections (if any) to the to-be-expanded transform, along
     * with their coders and windowing strategies.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.Components components = 1;
     */
    components?: Components;
    /**
     * The actual PTransform to be expaneded according to its spec.
     * Its input should be set, but its subtransforms and outputs
     * should not be.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.PTransform transform = 2;
     */
    transform?: PTransform;
    /**
     * A namespace (prefix) to use for the id of any newly created
     * components.
     *
     * @generated from protobuf field: string namespace = 3;
     */
    namespace: string;
}
// ...
class ExpansionRequest$Type extends MessageType<ExpansionRequest> {
    constructor() {
        super("org.apache.beam.model.expansion.v1.ExpansionRequest", [
            { no: 1, name: "components", kind: "message", T: () => Components },
            { no: 2, name: "transform", kind: "message", T: () => PTransform },
            { no: 3, name: "namespace", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<ExpansionRequest>): ExpansionRequest {
        const message = { namespace: "" };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<ExpansionRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ExpansionRequest): ExpansionRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* org.apache.beam.model.pipeline.v1.Components components */ 1:
                    message.components = Components.internalBinaryRead(reader, reader.uint32(), options, message.components);
                    break;
                case /* org.apache.beam.model.pipeline.v1.PTransform transform */ 2:
                    message.transform = PTransform.internalBinaryRead(reader, reader.uint32(), options, message.transform);
                    break;
                case /* string namespace */ 3:
                    message.namespace = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ExpansionRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* org.apache.beam.model.pipeline.v1.Components components = 1; */
        if (message.components)
            Components.internalBinaryWrite(message.components, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* org.apache.beam.model.pipeline.v1.PTransform transform = 2; */
        if (message.transform)
            PTransform.internalBinaryWrite(message.transform, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* string namespace = 3; */
        if (message.namespace !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.namespace);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message org.apache.beam.model.expansion.v1.ExpansionRequest
 */
export const ExpansionRequest = new ExpansionRequest$Type();