// sdks/node-ts/src/apache_beam/proto/beam_runner_api.ts
export interface Components {
    /**
     * (Required) A map from pipeline-scoped id to PTransform.
     *
     * Keys of the transforms map may be used by runners to identify pipeline
     * steps. Hence it's recommended to use strings that are not too long that
     * match regex '[A-Za-z0-9-_]+'.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.PTransform> transforms = 1;
     */
    transforms: {
        [key: string]: PTransform;
    };
    /**
     * (Required) A map from pipeline-scoped id to PCollection.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.PCollection> pcollections = 2;
     */
    pcollections: {
        [key: string]: PCollection;
    };
    /**
     * (Required) A map from pipeline-scoped id to WindowingStrategy.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.WindowingStrategy> windowing_strategies = 3;
     */
    windowingStrategies: {
        [key: string]: WindowingStrategy;
    };
    /**
     * (Required) A map from pipeline-scoped id to Coder.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.Coder> coders = 4;
     */
    coders: {
        [key: string]: Coder;
    };
    /**
     * (Required) A map from pipeline-scoped id to Environment.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.Environment> environments = 5;
     */
    environments: {
        [key: string]: Environment;
    };
}
// ...
export interface PTransform {
    /**
     * (Required) A unique name for the application node.
     *
     * Ideally, this should be stable over multiple evolutions of a pipeline
     * for the purposes of logging and associating pipeline state with a node,
     * etc.
     *
     * If it is not stable, then the runner decides what will happen. But, most
     * importantly, it must always be here and be unique, even if it is
     * autogenerated.
     *
     * @generated from protobuf field: string unique_name = 5;
     */
    uniqueName: string;
    /**
     * (Optional) A URN and payload that, together, fully defined the semantics
     * of this transform.
     *
     * If absent, this must be an "anonymous" composite transform.
     *
     * For primitive transform in the Runner API, this is required, and the
     * payloads are well-defined messages. When the URN indicates ParDo it
     * is a ParDoPayload, and so on. For some special composite transforms,
     * the payload is also officially defined. See StandardPTransforms for
     * details.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.FunctionSpec spec = 1;
     */
    spec?: FunctionSpec;
    /**
     * (Optional) A list of the ids of transforms that it contains.
     *
     * Primitive transforms (see StandardPTransforms.Primitives) are not allowed
     * to specify subtransforms.
     *
     * Note that a composite transform may have zero subtransforms as long as it
     * only outputs PCollections that are in its inputs.
     *
     * @generated from protobuf field: repeated string subtransforms = 2;
     */
    subtransforms: string[];
    /**
     * (Required) A map from local names of inputs (unique only with this map, and
     * likely embedded in the transform payload and serialized user code) to
     * PCollection ids.
     *
     * The payload for this transform may clarify the relationship of these
     * inputs. For example:
     *
     *  - for a Flatten transform they are merged
     *  - for a ParDo transform, some may be side inputs
     *
     * All inputs are recorded here so that the topological ordering of
     * the graph is consistent whether or not the payload is understood.
     *
     * @generated from protobuf field: map<string, string> inputs = 3;
     */
    inputs: {
        [key: string]: string;
    };
    /**
     * (Required) A map from local names of outputs (unique only within this map,
     * and likely embedded in the transform payload and serialized user code)
     * to PCollection ids.
     *
     * The URN or payload for this transform node may clarify the type and
     * relationship of these outputs. For example:
     *
     *  - for a ParDo transform, these are tags on PCollections, which will be
     *    embedded in the DoFn.
     *
     * @generated from protobuf field: map<string, string> outputs = 4;
     */
    outputs: {
        [key: string]: string;
    };
    /**
     * (Optional) Static display data for this PTransform application. If
     * there is none, it may be omitted.
     *
     * @generated from protobuf field: repeated org.apache.beam.model.pipeline.v1.DisplayData display_data = 6;
     */
    displayData: DisplayData[];
    /**
     * Environment where the current PTransform should be executed in.
     *
     * Transforms that are required to be implemented by a runner must omit this.
     * All other transforms are required to specify this.
     *
     * @generated from protobuf field: string environment_id = 7;
     */
    environmentId: string;
    /**
     * (Optional) A map from URNs designating a type of annotation, to the
     * annotation in binary format. For example, an annotation could indicate
     * that this PTransform has specific privacy properties.
     *
     * A runner MAY ignore types of annotations it doesn't understand. Therefore
     * annotations MUST NOT be used for metadata that can affect correct
     * execution of the transform.
     *
     * @generated from protobuf field: map<string, bytes> annotations = 8;
     */
    annotations: {
        [key: string]: Uint8Array;
    };
}
// ...
export enum IsBounded_Enum {
    /**
     * @generated from protobuf enum value: UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: UNBOUNDED = 1;
     */
    UNBOUNDED = 1,
    /**
     * @generated from protobuf enum value: BOUNDED = 2;
     */
    BOUNDED = 2
}
// ...
export interface Coder {
    /**
     * (Required) A specification for the coder, as a URN plus parameters. This
     * may be a cross-language agreed-upon format, or it may be a "custom coder"
     * that can only be used by a particular SDK. It does not include component
     * coders, as it is beneficial for these to be comprehensible to a runner
     * regardless of whether the binary format is agreed-upon.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.FunctionSpec spec = 1;
     */
    spec?: FunctionSpec;
    /**
     * (Optional) If this coder is parametric, such as ListCoder(VarIntCoder),
     * this is a list of the components. In order for encodings to be identical,
     * the FunctionSpec and all components must be identical, recursively.
     *
     * @generated from protobuf field: repeated string component_coder_ids = 2;
     */
    componentCoderIds: string[];
}
// ...
export enum MergeStatus_Enum {
    /**
     * @generated from protobuf enum value: UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * The WindowFn does not require merging.
     * Examples: global window, FixedWindows, SlidingWindows
     *
     * @generated from protobuf enum value: NON_MERGING = 1;
     */
    NON_MERGING = 1,
    /**
     * The WindowFn is merging and the PCollection has not had merging
     * performed.
     * Example: Sessions prior to a GroupByKey
     *
     * @generated from protobuf enum value: NEEDS_MERGE = 2;
     */
    NEEDS_MERGE = 2,
    /**
     * The WindowFn is merging and the PCollection has had merging occur
     * already.
     * Example: Sessions after a GroupByKey
     *
     * @generated from protobuf enum value: ALREADY_MERGED = 3;
     */
    ALREADY_MERGED = 3
}
/**
 * Whether or not subsequent outputs of aggregations should be entire
 * replacement values or just the aggregation of inputs received since
 * the prior output.
 *
 * @generated from protobuf message org.apache.beam.model.pipeline.v1.AccumulationMode
 */
export interface AccumulationMode {
}
/**
 * @generated from protobuf enum org.apache.beam.model.pipeline.v1.AccumulationMode.Enum
 */
export enum AccumulationMode_Enum {
    /**
     * @generated from protobuf enum value: UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * The aggregation is discarded when it is output
     *
     * @generated from protobuf enum value: DISCARDING = 1;
     */
    DISCARDING = 1,
    /**
     * The aggregation is accumulated across outputs
     *
     * @generated from protobuf enum value: ACCUMULATING = 2;
     */
    ACCUMULATING = 2,
    /**
     * The aggregation emits retractions when it is output
     *
     * @generated from protobuf enum value: RETRACTING = 3;
     */
    RETRACTING = 3
}
/**
 * Controls whether or not an aggregating transform should output data
 * when a window expires.
 *
 * @generated from protobuf message org.apache.beam.model.pipeline.v1.ClosingBehavior
 */
export interface ClosingBehavior {
}
/**
 * @generated from protobuf enum org.apache.beam.model.pipeline.v1.ClosingBehavior.Enum
 */
export enum ClosingBehavior_Enum {
    /**
     * @generated from protobuf enum value: UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * Emit output when a window expires, whether or not there has been
     * any new data since the last output.
     *
     * @generated from protobuf enum value: EMIT_ALWAYS = 1;
     */
    EMIT_ALWAYS = 1,
    /**
     * Only emit output when new data has arrives since the last output
     *
     * @generated from protobuf enum value: EMIT_IF_NONEMPTY = 2;
     */
    EMIT_IF_NONEMPTY = 2
}
/**
 * Controls whether or not an aggregating transform should output data
 * when an on-time pane is empty.
 *
 * @generated from protobuf message org.apache.beam.model.pipeline.v1.OnTimeBehavior
 */
export interface OnTimeBehavior {
}
/**
 * @generated from protobuf enum org.apache.beam.model.pipeline.v1.OnTimeBehavior.Enum
 */
export enum OnTimeBehavior_Enum {
    /**
     * @generated from protobuf enum value: UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * Always fire the on-time pane. Even if there is no new data since
     * the previous firing, an element will be produced.
     *
     * @generated from protobuf enum value: FIRE_ALWAYS = 1;
     */
    FIRE_ALWAYS = 1,
    /**
     * Only fire the on-time pane if there is new data since the previous firing.
     *
     * @generated from protobuf enum value: FIRE_IF_NONEMPTY = 2;
     */
    FIRE_IF_NONEMPTY = 2
}
/**
 * When a number of windowed, timestamped inputs are aggregated, the timestamp
 * for the resulting output.
 *
 * @generated from protobuf message org.apache.beam.model.pipeline.v1.OutputTime
 */
export interface OutputTime {
}
/**
 * @generated from protobuf enum org.apache.beam.model.pipeline.v1.OutputTime.Enum
 */
export enum OutputTime_Enum {
    /**
     * @generated from protobuf enum value: UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * The output has the timestamp of the end of the window.
     *
     * @generated from protobuf enum value: END_OF_WINDOW = 1;
     */
    END_OF_WINDOW = 1,
    /**
     * The output has the latest timestamp of the input elements since
     * the last output.
     *
     * @generated from protobuf enum value: LATEST_IN_PANE = 2;
     */
    LATEST_IN_PANE = 2,
    /**
     * The output has the earliest timestamp of the input elements since
     * the last output.
     *
     * @generated from protobuf enum value: EARLIEST_IN_PANE = 3;
     */
    EARLIEST_IN_PANE = 3
}
// ...
export interface Trigger_Default {
}
// ...
export interface FunctionSpec {
    /**
     * (Required) A URN that describes the accompanying payload.
     * For any URN that is not recognized (by whomever is inspecting
     * it) the parameter payload should be treated as opaque and
     * passed as-is.
     *
     * @generated from protobuf field: string urn = 1;
     */
    urn: string;
    /**
     * (Optional) The data specifying any parameters to the URN. If
     * the URN does not require any arguments, this may be omitted.
     *
     * @generated from protobuf field: bytes payload = 3;
     */
    payload: Uint8Array;
}
// ...
export interface DisplayData {
    /**
     * A key used to describe the type of display data. See StandardDisplayData
     * for the set of well known urns describing how the payload is meant to be
     * interpreted.
     *
     * @generated from protobuf field: string urn = 1;
     */
    urn: string;
    /**
     * (Optional) The data specifying any parameters to the URN. If
     * the URN does not require any arguments, this may be omitted.
     *
     * @generated from protobuf field: bytes payload = 2;
     */
    payload: Uint8Array;
}
// ...
export const Components = new Components$Type();
// ...
export const PTransform = new PTransform$Type();
// ...
export const PCollection = new PCollection$Type();
// ...
export const Coder = new Coder$Type();
// ...
export const WindowingStrategy = new WindowingStrategy$Type();
// ...
export const Trigger_Default = new Trigger_Default$Type();
// ...
export const Environment = new Environment$Type();
// ...
export const FunctionSpec = new FunctionSpec$Type();
// ...
export const DisplayData = new DisplayData$Type();